ltr_substr0(): Validate input used in 'test' algebra
Because 'test' likes to cry like a little child when it encounters
non-numeric argument. That aside, 'test' has the potential to hang the
shell execution when combined with a loop as it will never break it if
its argument is not numeric.

Well, we can in theory replace every instance of 'test' with a case
statement. Practically, we can just assert the input is numeric.

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>