esc_str(): Use explicit options
With the added benefit of easy parsing and extendable code. It is
important to note that this change also introduces undefined behavior;
that is, if both `-sqesc` and `-strip` are specified, the function may
not raise an error and will instead perform both stripping and escaping,
including a special escape on single quote characters.

Signed-off-by: Dimitar Yurukov <mscalindt@gmail.com>