pline(): Make `-nout` default behavior
I have come to the conclusion that `-nout` should be the default
behavior and no output should be printed. This generally aligns well
with advanced POSIX shell coding practices where the result of most
functions is stored in a variable and directly used, to maximize
performance.

In the rare instances where one might want to print the output of this
function, the shell provides a straightforward construct to safely print
the content of any variable. Thus, the output of this function can be
printed in the following manner:

`pline ...; printf "%s" "$_line"`

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>