mask_check(): Rework an incorrect variable value assert
In our case, `$1` will always be present after the `eval` call. So, the
assert is essentially incorrect in what it tried to assert. And, the
assert would have been superfluous even if it worked, because it doesn't
complement completely the argument fairness that `test` (the `[`
construct) expects.

To fix that, after the `eval` call, the following has to be done in the
order written:
(1) Assert $2 exists.
(2) Assert $1 is a whole number.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>