assert(): (documentation) provide reference for locale
POSIX.1-2017 states:
"In the POSIX locale, a range expression represents the set of collating
elements that fall between two elements in the collation sequence,
inclusive. In other locales, a range expression has unspecified
behavior: strictly conforming applications shall not rely on whether the
range expression is valid, or on the set of collating elements matched.
A range expression shall be expressed as the starting point and the
ending point separated by a <hyphen-minus> ( '-' )."

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>