arg_save(): (documentation) reference for eval
It is quite hard to provide a direct reference for the safe behavior
seen in our specific construction here, as it is a combination of
several fundamental principles and rules of the shell language.

Nonetheless, POSIX.1-2017 states:
- "A <backslash> that is not quoted shall preserve the literal value of
the following character, with the exception of a <newline>".
- "Enclosing characters in single-quotes ( '' ) shall preserve the
literal value of each character within the single-quotes.".
- "Enclosing characters in double-quotes ( "" ) shall preserve the
literal value of all characters within the double-quotes, with the
exception of the characters backquote, <dollar-sign>, and <backslash>".

The single quotes here are both literal and carry significance
throughout the word expansions and quoting, enabling the depicted safe
behavior of `eval` in our implementation.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>