ftype(): Improve `%/` PE with dynamic `/` match
By itself, `%/` makes perfect sense and serves its purpose well in
removing a trailing slash from the operand being tested. Trailing
slash(es) must not be present for operator `-h` of `test`, otherwise it
will cause an actual symbolic link to be followed and thus deteriorate
the reliability of `-h`.

However, `%/` will not suffice in removing more than one trailing slash.
The fix that costs nothing, is to make the PE dynamic by supplementing
it with a BRE, ensuring the removal of all trailing slashes.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>