assert(): [doc] define `-fs-canonical-path` further
As stated, a canonical path is a "largely implementation-defined
string". In our assertive implementation, whether a semantically
canonical path exists or not, is not taken into account.

A canonical path can be semantic but may not exist on the filesystem. It
makes sense to preserve this assumption. If ever needed, external
implementations can build upon `-fs-canonical-path` with a separate
assertive logic for existence.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>