assert(): [-fs-canonical-path] optimize loop
The old implementation will loop two times for the first iteration,
depending on whether the path ends with a trailing slash. This behavior
can be avoided. Additionally, the root directory (`/`) doesn't have to
be tested; it is implicitly specified by the POSIX standard that it is
always a true directory.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>