map_pchunk(): Utilize the custom concept of a path mask
With the concept of an LTR path mask, the function is free of any code
smell or cryptic shell code in its functionality. Additionally, the
LTR/RTL masks provide an efficient and complete filter over what is
processed.

Moving on, path_spec_static() will not be used internally and the path
will be processed as-is, using the state-of-the-art mask concept,
ensuring a standard, clear, and assertive mapping behavior.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>