map_pchunk_rev(): Fix bad PE strip of path segment
The parameter expansion (PE) used to strip the last path segment (and
consequently consume the path when only one path segment is remaining)
assumes a leading slash character will be present, resulting in a
NOP when it isn't. To be able to consume the path in all cases, it
should be modified to always strip a remainder instead.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>