map(): [doc] remove `.rc`
If the function itself does not return an error value, and `0` is
implied, `.rc` is not needed. Instead, any other sections like `.rc.fn`
or `.rc.util` shall serve as documentation.

The function itself is deemed to not return an error value if a shell
construct (except functions and utilities) is not being asserted with
`|| "$?"`.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>