ftype(): Always strip any trailing slashes if possible
The shell and its POSIX behavior introduce inherent ambiguity.

Although a trailing slash at the end of a path is a valid construct with
specific uses, it is a fragile and often unintended behavior that might
conflict with the purpose of ftype(). For robustness and reliability, as
part of a greater system, ftype() will reject it altogether.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>