ccount(): Allow multibyte and multiple characters
Since ccount() is not exactly aimed at POSIX compatibility as of
6cecf0ead690 [1] and 62b4c91c62a2 [2], further kind of undefined
behavior, such as the specification of multibyte characters, might be
appropriate.

Interestingly, with the restrictions lifted, it is possible for an
external implementation to build over the raw count in a "hacky" way to
produce the correct count; in `dash`, it appears at a glance that the
bytes of a character are respected, and is therefore only a matter of
division. That is, if the byte count and the interpreter are both known.

An interesting observation of implementation-defined behavior,
nonetheless.

[1] ccount(): Make IFS the default method
[2] src: Add ccount_posix()

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>