Revert "esc_sed(): [doc] note all explicit rc in `.rc`"
This reverts commit a3a69d7ee9c6214d2241534af84d8e00ecf7d295.

In favor of the following rule:
If `|| $?` is used on a shell construct or builtin, or if `return 0` is
not explicitly used and the function ends with shell code, it shall be
assumed and noted, that anything else besides rc `0` indicates `error`,
even alongside another meaning for specific return codes.

For different meaning of the specified return codes, which is made
possible due to the usage of functions (or utilities), `.rc.fn`
(or `.rc.util`) shall be consulted instead. Also applies to `.ec` and
related sections.