Revert "opt_fail(): [doc] add missing specifications for `error`"
This reverts commit c6eeeb704eb4df160103ffb4ba9480597efe2397.

In favor of the following rule:
If `|| $?` is used on a shell construct or builtin, or if `return 0` is
not explicitly used and the function ends with shell code, it shall be
assumed and noted, that anything else besides rc `0` indicates `error`,
even alongside another meaning for specific return codes.

For different meaning of the specified return codes, which is made
possible due to the usage of functions (or utilities), `.rc.fn`
(or `.rc.util`) shall be consulted instead. Also applies to `.ec` and
related sections.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>