path_spec_static(): Modularize and stabilize resolve behavior
This patch is both a refactor and a rework, to improve the modularity
and align the path resolving behavior closer to standards and/or
inherent logic, respectively.

The behavior is now clearly depicted and easy to follow, with added
documentation. The refactor allowed us to make the necessary changes
required for the proper handling of symbolic links (-h), and hence is
the behavior now stable and predictable as opposed to mixed and fragile.

Related changes that were necessary to make:
- e7d305ed9150 [1]. Might be reintroduced in the future.
- 10ce7a01b77a [2].
- d3a9ab4225bb [3].
- e28cda5691c2 [4].
- 790cded3a491 [5].

[1] path_spec_static(): Remove `-z` option
[2] path_spec_static(): Remove `-d`, `-e` support
[3] path_spec_static(): Simplify `-n` operation to single `/`
[4] path_spec_static(): Remove 'join path' support
[5] path_spec_static(): Remove `-f`, `-g`, `-j` support

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>