scopt(): Assert proper function input
Because the function does not deal with short/long variants of options,
we cannot assert input as precisely as in copt() or sopt().

But, we can assert all arguments have been parsed when the loop
naturally breaks or—the point of the input assert—when it breaks due
to empty input.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>