libfile_n(): Fix EOF additions parsing flaw
What allows us to consistently separate data from actual operations (so
as to not interpret any data of an operation as an operation) is the
matching logic of at least two arguments in our pseudo array, thus
making it impossible for the data to imply an operation.

An argument to cling on is not present for EOF. To fix the flaw, use a
dummy argument `0` when we parse it for our pseudo array and adjust the
matching logic accordingly.

Fixes: 2682b0afe16b [1].

[1] libfile_n(): Add support for EOF additions

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>