libfile_n(): Fix fd leak on error
Yeah, the sugar coated syntax looked too good to be true.

But even if the statement gets rewritten with the OR operator so that
the fd is always released, any original errors will be masked.
Therefore, when it comes to cleanup with fds, explicit error handling is
a requirement for proper error handling *and* cleanup.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>