fed(): Apply modifiers concept over default operations
With modifiers, miscellaneous (and default) behavior is documentable,
modular, and powerful. The usage of modifiers has also exposed fragile
interplay between the operations, which ultimately allowed us to spot
and fix inconsistent and confusing behavior in (current and future)
implementations.

Future extensions over the default operations will have to conform to
the concept of modifiers—which is a great safety net for certain
pitfalls.

Thus, a showcase of the new specification: {
  -m 37 -a ''      -> Add new line `` at "37".
  -m 37 -a foo     -> Add new line `foo` at "37".
  -a ''            -> <noop>
  -a foo           -> Add/suffix `foo` to line(s).

  -m 37 -r ''      -> Remove line "37".
  -m 37 -r foo     -> Remove line "37" if any string (lof) `foo`.
  -r ''            -> Remove line(s).
  -r foo           -> Remove line(s) if any string (lof) `foo`.

  -m 37 -s '' ''   -> Substitute line "37" with ``.
  -m 37 -s bar ''  -> Substitute line "37" with `` if (lof) `bar`.
  -s '' ''         -> Substitute line(s) with ``.
  -s bar ''        -> Substitute line(s) with `` if (lof) `bar`.
}

For clarity, <noop> is a valid operation that marks the line processed.
It shall only take place in inherent additions where nothing is changed.

Also, our DSL now mirrors the declarative, deterministic, and modular
design philosophy behind RFC 6902, which further reinforces the flawless
DSL behavior.

Signed-off-by: Dimitar Yurukov <mscalindt@protonmail.com>